refactor to classes - FakeSchema
