+++ src/fake-schema/fake-schema.test.ts
@@ -1,0 +1,22 @@
+import { fakeSchema } from "./";
+import { createSchemaApi } from "../build-schema";
+
+const idl = `
+  type Person {
+    name: String
+  }
+`;
+
+describe("fakeSchema", () => {
+  const readIDL = async () => idl;
+  const { buildServerSchema } = createSchemaApi({ readIDL });
+  const schema = buildServerSchema({ idl });
+
+  describe("schema and no config", () => {
+    const fakedSchema = fakeSchema(schema);
+
+    test("schema", () => {
+      expect(fakedSchema).toBeDefined();
+    });
+  });
+});
