+++ src/fake-schema/index.ts
@@ -1,0 +1,11 @@
+import { GraphQLSchema } from "graphql";
+import { FakeSchema } from "./FakeSchema";
+
+export function fakeSchema(
+  schema: GraphQLSchema,
+  config: any = {}
+): GraphQLSchema {
+  return new FakeSchema(schema, config).schema;
+}
+
+export { FakeSchema } from "./FakeSchema";
